+++ src/utils/settings.py
@@ -111,7 +111,7 @@ APPLICATION = "commix"
 DESCRIPTION_FULL = "Automated All-in-One OS Command Injection and Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
-VERSION_NUM = "2.4.7"
+VERSION_NUM = "2.4.8"
 STABLE_VERSION = False
 if STABLE_VERSION:
   VERSION = "v" + VERSION_NUM[:3] + "-stable"
@@ -678,7 +678,8 @@ TAMPER_SCRIPTS = {
                   "space2plus": False,
                   "space2htab": False,
                   "space2vtab": False,
-                  "singlequotes": False
+                  "singlequotes": False,
+                  "caret": False
                  }
 
 # HTTP Errors