+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 2.4 (upcoming)
+* Added: New tamper script "caret.py" that adds the caret symbol (^) between the characters of the generated payloads (for windows targets).
 * Added: New tamper script "singlequotes.py" that adds single quotes (') between the characters of the generated payloads (for *nix targets).
 
 ## Version 2.3 (2018-03-07)